Update to 16 bits with obstacles

+++ common/includes/Paquet/PaquetObstacle.hh
@@ -9,8 +9,8 @@ class		PaquetObstacle : public Paquet
 private:
   uint8_t	_id;
   uint8_t _type;
-  uint8_t _x;
-  uint8_t _y;
+  uint16_t _x;
+  uint16_t _y;
 
 public:
   PaquetObstacle();
@@ -20,11 +20,11 @@ public:
   virtual ~PaquetObstacle();
 
   void setType(const uint8_t&);
-  void setPosition(const uint8_t&, const uint8_t&);
+  void setPosition(const uint16_t&, const uint16_t&);
 
   const uint8_t& getType() const;
-  const uint8_t& getX() const;
-  const uint8_t& getY() const;
+  const uint16_t& getX() const;
+  const uint16_t& getY() const;
 
   void		createPaquet();
   void		parsePaquet();
